+++ test/command.js
@@ -484,4 +484,28 @@ describe('command run', function() {
             assert.equal(output, '1.2.3');
         });
     });
+
+    it('extend()', () => {
+        const invocations = [];
+        const extension = (...args) => {
+            invocations.push(args);
+        };
+
+        const command = cli
+            .command('test')
+            .extend(extension, 1, 2);
+        const nested = command
+            .command('nested')
+            .extend(extension)
+            .extend(extension, 1, 2, 3, 4);
+
+        command.extend(extension, 2, 3);
+
+        assert.deepEqual(invocations, [
+            [command, 1, 2],
+            [nested],
+            [nested, 1, 2, 3, 4],
+            [command, 2, 3]
+        ]);
+    });
 });